+++ frontend/fingertips-frontend/components/views/OneAreaView/index.tsx
@@ -1,10 +1,14 @@
 import { H2 } from 'govuk-react';
-import { chartOptions, getChartList, IViewProps } from '@/lib/viewUtils';
+import {
+  chartOptions,
+  getChartListForView,
+  IViewProps,
+} from '@/lib/viewUtils/viewUtils';
 
 export function OneAreaView({ areaCodes, indicatorsSelected }: IViewProps) {
   // based on the number of indicators
   // determine the visalisations to show
-  const chartList = getChartList(indicatorsSelected, 'oneAreaView');
+  const chartList = getChartListForView(indicatorsSelected, 'oneAreaView');
   // determine the data to fetch
 
   return (
