+++ frontend/fingertips-frontend/components/views/ThreeOrMoreAreasView/index.tsx
@@ -1,11 +1,14 @@
-import { getChartList, IViewProps } from '@/lib/viewUtils';
+import { getChartListForView, IViewProps } from '@/lib/viewUtils/viewUtils';
 import { H2 } from 'govuk-react';
 
 export function ThreeOrMoreAreasView({
   areaCodes,
   indicatorsSelected,
 }: IViewProps) {
-  const chartList = getChartList(indicatorsSelected, 'threeOrMoreAreasView');
+  const chartList = getChartListForView(
+    indicatorsSelected,
+    'threeOrMoreAreasView'
+  );
   return (
     <>
       <H2>View data for all selected areas</H2>
