+++ frontend/fingertips-frontend/components/views/TwoAreasView/index.tsx
@@ -1,8 +1,8 @@
-import { getChartList, IViewProps } from '@/lib/viewUtils';
+import { getChartListForView, IViewProps } from '@/lib/viewUtils/viewUtils';
 import { H2 } from 'govuk-react';
 
 export function TwoAreasView({ areaCodes, indicatorsSelected }: IViewProps) {
-  const chartList = getChartList(indicatorsSelected, 'twoAreasView');
+  const chartList = getChartListForView(indicatorsSelected, 'twoAreasView');
 
   return (
     <>
